+ push de control

See merge request groupf4294046/erp!11

+++ src/main/java/com/pet_it/program/DAO/employeeDAO.java
@@ -11,6 +11,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
  *
  * @author Jordan
  */
-public interface employeeDAO extends JpaRepository<employees, Integer> {
+public interface employeeDAO extends JpaRepository<employees, Long> {
 
 }
